Date/StatusGroup forms + *major refactor*
CHANGE 1
-------------
Now all xxxxGroupEditBox are able to interact directly with the model
classes. This avoid endless duplication of code such as:
    m_editTypeGroup->editTypeLabel()->text() etc...

1) The constructor receives a pointer which is used for the
inizialitation of data fields;
2) The xxxxGroup() method returns a new xxxxGroup initialized with the
values in the form;
3) The updateExistingxxxxGroup(xxxxGroup *) receives an object of type
xxxxGroup and sets its internal field according to the values in the
form. This version is used every time we already have an object that
inherits from xxxxGroup.

See one of the updated forms for details.

CHANGE 2 (uncomplete)
-------------
Every time there is an optional field we should grant the user with the
possibility of adding and removing it. In particular we should be able
to discriminate between an unsetted and a setted field.
In the case of QLineEdits is easy. An empty string is translated in an
absent attribute.
But what with QDateEdits, QSpinBoxes, etc. which *always* contain a
valid value/date?
My proposed solution (see DateGroupEditBox) is to associate a QCheckBox
to every problematic field. This way the user can manually select which
attributes include.

I had also to change model/dategroup.* because it was impossible to
remove the two optional year (int) fields from a created object. Other
classes in the model have similar issues. Be advised and fix them while
you introduce QCheckBoxes.