Add code to insert formats in the formatMap
Also fix SIGSEGV caused by PublicationType taking ownership of
FormatType pointers.

Another bug has arised though, in the UI the publication channel is not
optional. See you at next commit.

Note: Format are removed from formatMap in the deleteClicked() method of
CoreMetadataTypeForm. The insertion in formatMap is handled by the
applyClicked() method of FormatTypeForm. This is the only place where we
know which was the old ID and where we are able to remove it from the
formatMap.
Also I have added a validity check to avoid duplications of format IDs.